refactor(e2e): commonize test server setup for client-js e2e tests
- Create createTestServerSetup factory function in _test-utils
- Fix server leak by using proper async cleanup with promisified server.close()
- Server is now created inside setup function with try/catch for cleanup on error
- Simplify zod-v3 and zod-v4 setup.ts to use shared factory
- Add necessary dependencies to _test-utils package.json
