+++ e2e-tests/client-js/_test-utils/package.json
@@ -8,14 +8,25 @@
   "exports": {
     ".": "./src/index.ts"
   },
+  "dependencies": {
+    "@hono/node-server": "^1.14.3",
+    "hono": "^4.6.0",
+    "get-port": "^7.1.0"
+  },
   "peerDependencies": {
     "@mastra/client-js": "*",
     "@mastra/core": "*",
+    "@mastra/hono": "*",
+    "@mastra/libsql": "*",
+    "@mastra/observability": "*",
     "vitest": "*"
   },
   "devDependencies": {
     "@mastra/client-js": "workspace:*",
     "@mastra/core": "workspace:*",
+    "@mastra/hono": "workspace:*",
+    "@mastra/libsql": "workspace:*",
+    "@mastra/observability": "workspace:*",
     "vitest": "catalog:",
     "typescript": "catalog:"
   }
